car_picker/requirements.txt
streamlit>=1.28
pillow>=10.0
streamlit-autorefresh>=1.0.1
